==> python_SouthAmerica/4_Coregistration.ipynb <==
# import Utils

# #
# # SEE HERE! 
# # This cell must run first to setup the working environment
# #

# #load the config file
# Utils.load_config('config_NA.json')
# Utils.print_parameters()

# #let's print out some important parameters
# #you can change the 'input_file' in config.json to use different mineral deposits. 
# #Remember the files we have created in step 2?
# print('The file name of the mineral deposits: ', Utils.get_parameter('coreg_input_files'))
# print('The output folder: ', Utils.get_coreg_input_dir())
# print('The region of interest(in degree): ', Utils.get_parameter('regions'))
# print('The subduction convergence kinematics statistics file name template: ', Utils.get_parameter('vector_files'))
# print('\n')

# import coregistration
# coregistration.run() #run the coregistration script
# #some files should have been created at this point
# #let's move to the next cell and check the result

import parameters as parameters
import Utils

Utils.load_config('config_SA.json')
Utils.print_parameters()

#let's print out some important parameters
#you can change the 'input_file' in config.json to use different mineral deposits. 
#Remember the files we have created in step 2?
print('The file name of the mineral deposits: ', Utils.get_parameter('coreg_input_files'))
print('The output folder: ', Utils.get_coreg_input_dir())
print('The region of interest(in degree): ', Utils.get_parameter('regions'))
print('The subduction convergence kinematics statistics file name template: ', Utils.get_parameter('vector_files'))
print('\n')

import coregistration
coregistration.run() #run the coregistration script
#some files should have been created at this point
#let's move to the next cell and check the results
