==> python_SouthAmerica/6_Age_Cleaning.ipynb <==
import csv
import numpy as np

# #NA_positive_data = pd.read_csv('case_AREPS/coreg_output_clean/NA/positive_deposits.csv')

# filename = 'case_AREPS/coreg_output_clean/NA/positive_deposits.csv'

# lats, lons, minage = [], [], []

# with open(filename) as f:
#     reader = csv.reader(f)
#     next(reader)
#     for row in reader:
#         lats.append(float(row[1]))
#         lons.append(float(row[0]))
#         minage.append(float(row[2]))

        
# NA_occurrences = []
# NA_dep_ages = []
        
# NA_occurrences = np.bincount(minage)

# for n in range(1, len(NA_occurrences)):
#     if NA_occurrences[n] > 0:
#         NA_dep_ages.append(n)

# print('For North America:')
# #print(non_dep_ages)
# NA_dep_ages = np.array(NA_dep_ages)
# print(NA_dep_ages)
# print(len(NA_dep_ages))
# #print(len(non_dep_ages))
# print('')

filename = 'case_AREPS/coreg_output_clean/SA/positive_deposits.csv'

lats, lons, minage = [], [], []

with open(filename) as f:
    reader = csv.reader(f)
    next(reader)
    for row in reader:
        lats.append(float(row[1]))
        lons.append(float(row[0]))
        minage.append(float(row[2]))


SA_occurrences = []
SA_dep_ages = []
        
SA_occurrences = np.bincount(minage)

for n in range(1, len(SA_occurrences)):
    if SA_occurrences[n] > 0:
        SA_dep_ages.append(n)
        
print('For South America:')
#print(SA_non_dep_ages)
SA_dep_ages = np.array(SA_dep_ages)
print(SA_dep_ages)
print(len(SA_dep_ages))
#print(len(non_dep_ages))
#print((occurrences))


# importing pandas module 
import pandas as pd 

#NA_volcanics = 'case_AREPS/coreg_output_clean/NA/negative_deposits.csv'
SA_volcanics = 'case_AREPS/coreg_output_clean/SA/negative_deposits.csv'


# making data frame from csv file 
#NA_data = pd.read_csv(NA_volcanics, index_col ="age" )
SA_data = pd.read_csv(SA_volcanics, index_col ="age" ) 

# print('Initial number of data points for North America: ')
# print(NA_data.shape[0])
print('Initial number of data points for South America: ')
print(SA_data.shape[0])

# dropping passed values 
# NA_data.drop(NA_dep_ages, inplace=True, errors='ignore') 
SA_data.drop(SA_dep_ages, inplace=True, errors='ignore')

# print('Final number of data points for North America: ')
# print(NA_data.shape[0])
print('Final number of data points for South America: ')
print(SA_data.shape[0])

# NA_data.to_csv('case_AREPS/coreg_output_clean/NA/negative_deposits_cleaned.csv', index=True)
SA_data.to_csv('case_AREPS/coreg_output_clean/SA/negative_deposits_cleaned.csv', index=True)
#print(non_dep_ages)
#random_ages_nondep = np.array(non_dep_ages)

import pandas as pd

#NA_positive_data = pd.read_csv('case_AREPS/coreg_output_clean/NA/positive_deposits.csv')
#NA_negative_data = pd.read_csv('case_AREPS/coreg_output_clean/NA/negative_deposits_cleaned.csv')

SA_positive_data = pd.read_csv('case_AREPS/coreg_output_clean/SA/positive_deposits.csv')
SA_negative_data = pd.read_csv('case_AREPS/coreg_output_clean/SA/negative_deposits_cleaned.csv')



#NA_positive_data['label']='NA_Porphyries'
#NA_negative_data['label']='NA_Volcanics'
SA_positive_data['label']='SA_Porphyries'
SA_negative_data['label']='SA_Volcanics'

# frames = [NA_positive_data, NA_negative_data, SA_positive_data, SA_negative_data]
frames = [SA_positive_data, SA_negative_data]

result = pd.concat(frames)

result.to_csv('case_AREPS/coreg_output_clean/all_data_iogas.csv', index=True)
