==> python_SouthAmerica/7_Data_Wrangling.ipynb <==
##### After having a look at the data, let's start selecting features and add labels.

import numpy as np
import pandas as pd
from parameters import parameters 
import Utils as Utils
import os

Utils.create_dirs()


#coreg_out_dir = Utils.get_coreg_output_dir()

# NA_positive_data = pd.read_csv('case_AREPS/coreg_output_clean/NA/positive_deposits.csv')
# NA_negative_data = pd.read_csv('case_AREPS/coreg_output_clean/NA/negative_deposits_cleaned.csv')
# NA_candidates_data = pd.read_csv('case_AREPS/coreg_output_clean/NA/deposit_candidates.csv')
SA_positive_data = pd.read_csv('case_AREPS/coreg_output_clean/SA/positive_deposits.csv')
SA_negative_data = pd.read_csv('case_AREPS/coreg_output_clean/SA/negative_deposits_cleaned.csv')
SA_candidates_data = pd.read_csv('case_AREPS/coreg_output_clean/SA/deposit_candidates.csv')


feature_names = parameters['feature_names']

# NA_positive_feat = NA_positive_data[feature_names].dropna()
# NA_negative_feat = NA_negative_data[feature_names].dropna()
# NA_candidates_feat = NA_candidates_data[feature_names].dropna()
SA_positive_feat = SA_positive_data[feature_names].dropna()
SA_negative_feat = SA_negative_data[feature_names].dropna()
SA_candidates_feat = SA_candidates_data[feature_names].dropna()


# NA_positive_feat['label']=True
# NA_negative_feat['label']=False
SA_positive_feat['label']=True
SA_negative_feat['label']=False


#save the data


# NA_positive_feat.to_csv(Utils.get_ml_input_dir() + 'NA_positive_deposits.csv', index=False)
# NA_negative_feat.to_csv(Utils.get_ml_input_dir() + 'NA_negative_deposits.csv', index=False)
# NA_candidates_feat.to_csv(Utils.get_ml_input_dir() + 'NA_deposit_candidates.csv', index=False)
SA_positive_feat.to_csv(Utils.get_ml_input_dir() + 'SA_positive_deposits.csv', index=False)
SA_negative_feat.to_csv(Utils.get_ml_input_dir() + 'SA_negative_deposits.csv', index=False)
SA_candidates_feat.to_csv(Utils.get_ml_input_dir() + 'SA_deposit_candidates.csv', index=False)



# NA_positive_data.iloc[NA_positive_feat.index].to_csv(Utils.get_ml_input_dir() + 'NA_positive_all_columns.csv', index=False)
# NA_negative_data.iloc[NA_negative_feat.index].to_csv(Utils.get_ml_input_dir() + 'NA_negative_all_columns.csv', index=False)
# NA_candidates_data.iloc[NA_candidates_feat.index].to_csv(Utils.get_ml_input_dir() + 'NA_candidates_all_columns.csv', index=False)
SA_positive_data.iloc[SA_positive_feat.index].to_csv(Utils.get_ml_input_dir() + 'SA_positive_all_columns.csv', index=False)
SA_negative_data.iloc[SA_negative_feat.index].to_csv(Utils.get_ml_input_dir() + 'SA_negative_all_columns.csv', index=False)
SA_candidates_data.iloc[SA_candidates_feat.index].to_csv(Utils.get_ml_input_dir() + 'SA_candidates_all_columns.csv', index=False)


import glob
files = glob.glob(Utils.get_ml_input_dir() + '*')
print('\ngenerated files:')
for f in files:
    print(f)


import pandas as pd

# NA_positive_data = pd.read_csv('case_AREPS/ml_input/NA_positive_all_columns.csv')
# NA_negative_data = pd.read_csv('case_AREPS/ml_input/NA_negative_all_columns.csv')

SA_positive_data = pd.read_csv('case_AREPS/ml_input/SA_positive_all_columns.csv')
SA_negative_data = pd.read_csv('case_AREPS/ml_input/SA_negative_all_columns.csv')



# NA_positive_data['label']='NA_Porphyries'
# NA_negative_data['label']='NA_Volcanics'
SA_positive_data['label']='SA_Porphyries'
SA_negative_data['label']='SA_Volcanics'

# frames = [NA_positive_data, NA_negative_data, SA_positive_data, SA_negative_data]
frames = [SA_positive_data, SA_negative_data]
result = pd.concat(frames)

result.to_csv('case_AREPS/ml_input/all_data_iogas.csv', index=True)
