==> python_SouthAmerica/9_ML_Export_Data_MEAN.ipynb <==
import csv
import numpy as np
import pandas as pd


# For South America


## Paleocene to early Eocene belt (67 to 47 Ma)##

#age_min = str(mineral)
met_age_min = 47 
met_age_max = 67

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_Neural_Network.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_MLP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_SVC_RBF.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_RBF_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_GaussianProcess.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_GP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)


## Middle Eocene to early Oligocene belt (47 to 27 Ma)##

#age_min = str(mineral)
met_age_min = 27 
met_age_max = 48

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_Neural_Network.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_MLP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_SVC_RBF.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_RBF_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_GaussianProcess.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_GP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

## Miocene to early Pliocene belt (23 to 3 Ma)##

#age_min = str(mineral)
met_age_min = 3 
met_age_max = 24

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_Neural_Network.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_MLP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_SVC_RBF.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_RBF_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_GaussianProcess.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_GP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

## Cenozoic (68 to 3 Ma)##

#age_min = str(mineral)
met_age_min = 3 
met_age_max = 67

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_Neural_Network.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_MLP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_SVC_RBF.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_RBF_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)

SA_candidates_prob = pd.read_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_GaussianProcess.csv')
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age >= met_age_min]
SA_candidates_prob = SA_candidates_prob[SA_candidates_prob.age < met_age_max]
SA_candidates_prob = SA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
SA_candidates_prob.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_prob_GP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)



import csv
import numpy as np
import pandas as pd


# # For North America

# for mineral in range(0,231,10):
#     age_min = str(mineral)
#     met_age_min = mineral 
#     met_age_max = mineral+11
#     NA_candidates_prob = pd.read_csv('case_AREPS/ml_output/NA/NA_candidates_all_prob_GaussianProcess.csv')
#     NA_candidates_prob = NA_candidates_prob[NA_candidates_prob.age >= met_age_min]
#     NA_candidates_prob = NA_candidates_prob[NA_candidates_prob.age < met_age_max]
#     NA_candidates_prob = NA_candidates_prob.groupby(['lon','lat'])['prob'].mean()
#     NA_candidates_prob.to_csv('case_AREPS/ml_output/NA/by_age/NA_candidates_all_prob_GP_'+str(met_age_min)+'_'+str(met_age_max)+'.csv', index=True)


import pandas as pd
from pathlib import Path

source_files = sorted(Path('case_AREPS/ml_output/SA/by_age/').glob('*.csv'))

dataframes = []
for file in source_files:
    df = pd.read_csv(file) # additional arguments up to your need
    df['classifier'] = file.name
    dataframes.append(df)

SA_by_age_merged = pd.concat(dataframes)
SA_by_age_merged.to_csv('case_AREPS/ml_output/SA/by_age/SA_candidates_all_files_merged.csv', index=True)

# import pandas as pd
# from pathlib import Path

# source_files = sorted(Path('case_AREPS/ml_output/NA/by_age/').glob('*.csv'))

# dataframes = []
# for file in source_files:
#     df = pd.read_csv(file) # additional arguments up to your need
#     df['classifier'] = file.name
#     dataframes.append(df)

# NA_by_age_merged = pd.concat(dataframes)
# NA_by_age_merged.to_csv('case_AREPS/ml_output/NA/by_age/NA_candidates_all_files_merged.csv', index=True)
