==> python_SouthAmerica/9_ML_Export_Data_SA.ipynb <==
from parameters import parameters
import numpy as np
import pandas as pd
import Utils
#
# LOOK HERE! 
# This cell must run first to setup the working environment
#

#load the config file
Utils.load_config('config.json')
Utils.print_parameters()

import numpy as np
import matplotlib.pyplot as plt
from matplotlib.colors import ListedColormap
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.datasets import make_moons, make_circles, make_classification
from sklearn.neural_network import MLPClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC
from sklearn.gaussian_process import GaussianProcessClassifier
from sklearn.metrics import plot_confusion_matrix
from sklearn.gaussian_process.kernels import RBF
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from sklearn import preprocessing
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import ShuffleSplit
from sklearn.model_selection import KFold
import pandas as pd
from parameters import parameters 
import Utils as Utils

names = ["GaussianProcess", "Neural_Network", "SVC_POLY", "SVC_RBF", "AdaBoost", "Random_Forest",
        "Decision_Tree","QDA", "Naive_Bayes"]

classifiers = [
    GaussianProcessClassifier(),
    MLPClassifier(alpha=0.1, max_iter=10000),
    SVC(kernel='poly', degree=3, gamma=2, C=10, probability=True),
    SVC(kernel='rbf', gamma=2, C=10, probability=True),
    AdaBoostClassifier(n_estimators=10),
    RandomForestClassifier(max_depth=5, n_estimators=10, max_features=1),
    DecisionTreeClassifier(max_depth=5),
    QuadraticDiscriminantAnalysis(),
    GaussianNB()]


#NA_positive = pd.read_csv(Utils.get_ml_input_dir() + 'NA_positive_deposits.csv')
#NA_negative = pd.read_csv(Utils.get_ml_input_dir() + 'NA_negative_deposits.csv')
SA_positive = pd.read_csv(Utils.get_ml_input_dir() + 'SA_positive_deposits.csv')
SA_negative = pd.read_csv(Utils.get_ml_input_dir() + 'SA_negative_deposits.csv')

#NA_train_test_data = pd.concat([NA_positive, NA_negative])
SA_train_test_data = pd.concat([SA_positive, SA_negative])

#NA_labels = NA_train_test_data.iloc[:,-1]
SA_labels = SA_train_test_data.iloc[:,-1]
#NA_data = NA_train_test_data.iloc[:,:-1]
SA_data = SA_train_test_data.iloc[:,:-1]

#NA_data = preprocessing.scale(NA_data)
SA_data = preprocessing.scale(SA_data)

# rs = ShuffleSplit(n_splits=10, test_size=.1, random_state=None)
# #rs = KFold(n_splits=10, shuffle=True, random_state=1)

# NA_all_scores_cv = []
# for name, clf in zip(names, classifiers):
#     scores = cross_val_score(clf, NA_data, NA_labels, cv=rs)
# #     score_list = scores.tolist()
# #     NA_all_scores_cv.append(name)
# #     NA_all_scores_cv.append(score_list)
#     print(name)
#     print(f"Score Mean: {round(np.mean(scores), 2)}, Score Standard Deviation: {round(np.std(scores),4)}")
#     print(' ')

# #np.savetxt('case_AREPS/ml_output/NA_cv_scores_ShuffleSplit_10_2.csv', NA_all_scores_cv, fmt='%s\n', delimiter=",")

# print('done!')

rs = ShuffleSplit(n_splits=10, test_size=.1, random_state=None)
#rs = KFold(n_splits=10, shuffle=True, random_state=1)

SA_all_scores_cv = []
for name, clf in zip(names, classifiers):  
    scores = cross_val_score(clf, SA_data, SA_labels, cv=rs)
#    score_list = scores.tolist()
#    SA_all_scores_cv.append(name)
#    SA_all_scores_cv.append(score_list)
    print(name)
    print(f"Score Mean: {round(np.mean(scores), 2)}, Score Standard Deviation: {round(np.std(scores),4)}")
    print(' ')
        
#np.savetxt('case_AREPS/ml_output/SA_cv_scores_ShuffleSplit_5.csv', SA_all_scores_cv, fmt='%s\n', delimiter=",")
print('done!')

%matplotlib inline

import csv
import numpy as np
import pandas as pd
import re
import matplotlib.pyplot as plt
import cartopy.crs as ccrs
import matplotlib.ticker as mticker
from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
from netCDF4 import Dataset
import TopoMap
import os
from scipy.interpolate import griddata
from scipy import misc, ndimage

#NA_positive = pd.read_csv(Utils.get_ml_input_dir() + 'NA_positive_deposits.csv')
#NA_negative = pd.read_csv(Utils.get_ml_input_dir() + 'NA_negative_deposits.csv')
SA_positive = pd.read_csv(Utils.get_ml_input_dir() + 'SA_positive_deposits.csv')
SA_negative = pd.read_csv(Utils.get_ml_input_dir() + 'SA_negative_deposits.csv')

#NA_train_test_data = pd.concat([NA_positive, NA_negative])
SA_train_test_data = pd.concat([SA_positive, SA_negative])

#NA_labels = NA_train_test_data.iloc[:,-1]
SA_labels = SA_train_test_data.iloc[:,-1]
#NA_data = NA_train_test_data.iloc[:,:-1]
SA_data = SA_train_test_data.iloc[:,:-1]

#NA_data = preprocessing.scale(NA_data)
SA_data = preprocessing.scale(SA_data)

#NA_train_data, NA_test_data, NA_train_labels, NA_test_labels = train_test_split(NA_data, NA_labels, test_size=0.4, random_state=None)

df = pd.DataFrame(columns=['lat', 'lon', 'recon_lon', 'recon_lat', 'age', 'prob', 'method']) 

for name, clf in zip(names, classifiers):
    print(name)
    
    clf.fit(SA_data, SA_labels)
    
    SA_candidates= pd.read_csv(Utils.get_ml_input_dir() + 'SA_deposit_candidates.csv')
    
    SA_all_candidates_data = pd.read_csv('case_AREPS/ml_input/SA_candidates_all_columns.csv')
    SA_all_positive_data = pd.read_csv('case_AREPS/ml_input/SA_positive_all_columns.csv')
    SA_all_negative_data = pd.read_csv('case_AREPS/ml_input/SA_negative_all_columns.csv')
    #NA_all_candidates_data = pd.read_csv('case_AREPS/coreg_output_clean/NA/deposit_candidates.csv', index_col ="age" )

    feature_names = Utils.get_parameter('feature_names')

    SA_positive_idx = SA_all_positive_data[feature_names].dropna().index
    SA_negative_idx = SA_all_negative_data[feature_names].dropna().index
    SA_candidates_idx = SA_all_candidates_data[feature_names].dropna().index
    
    SA_test_data = preprocessing.scale(SA_candidates)
    SA_mesh_prob=clf.predict_proba(SA_test_data)

    SA_candidates_lat_lon = SA_all_candidates_data.iloc[SA_candidates_idx][['lat', 'lon', 'recon_lon', 'recon_lat', 'age']]
    print(SA_candidates_lat_lon.shape) ###1  
    SA_candidates_lat_lon['prob'] = SA_mesh_prob[:,1]
    print(SA_candidates_lat_lon.shape)
    SA_candidates_lat_lon['method']=name
    print(SA_candidates_lat_lon.shape)
    df = pd.concat([df, SA_candidates_lat_lon]) 
    SA_candidates_lat_lon.to_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_'+name+'.csv', index=True)
    
    
# #     print(NA_candidates_lat_lon.shape) ###3
    
# #     NA_candidates_lat_lon = NA_candidates_lat_lon[NA_candidates_lat_lon.age > met_age_min]
# #     NA_candidates_lat_lon = NA_candidates_lat_lon[NA_candidates_lat_lon.age < met_age_max]
# # #    print(NA_candidates_lat_lon.shape) ###2 
    

# # #    print(NA_candidates_lat_lon.shape) ###2 
    
#     NA_candidates_lat_lon_group = NA_candidates_lat_lon.groupby(['recon_lon','recon_lat'])['prob'].mean()
#     print(NA_candidates_lat_lon_group.shape) ###4
#     NA_candidates_lat_lon_group.to_csv('case_AREPS/ml_output/NA/NA_candidates_lat_lon_all_ages_'+name+'.csv', index=True)


#    SA_candidates_lat_lon_group = SA_candidates_lat_lon.groupby(['lon','lat'])['prob'].mean()
#    print(NA_candidates_lat_lon_group.shape) ###4
#    SA_candidates_lat_lon_group.to_csv('case_AREPS/ml_output/SA/SA_candidates_all_mean_prob_'+name+'.csv', index=True)
    
        
# #     NA_candidates_lat_lon = NA_candidates_lat_lon.reset_index()
# #     print(NA_candidates_lat_lon.shape) ###5
    
# # #    NA_candidates_lat_lon.to_csv('case_AREPS/ml_output/NA/NA_candidates_mean_all_ages_'+name+'.csv', index=True)
#     NA_candidates_lat_lon.to_csv('case_AREPS/ml_output/NA/NA_candidates_recon_lat_lon_age_'+name+'.csv', index=True)


#df.to_csv('case_AREPS/ml_output/SA/SA_candidates_all_prob_all_classifers.csv', index=True) #Export all data, better go down and export this + positives and Negatives
    

# all_prob_adaboost = pd.read_csv('case_AREPS/ml_output/NA/NA_candidates_all_prob_AdaBoost-10.csv')
# all_prob_GaussianProcess = pd.read_csv('case_AREPS/ml_output/NA/NA_candidates_all_prob_GaussianProcess-1.csv')
# all_prob_NN = pd.read_csv('case_AREPS/ml_output/NA/NA_candidates_all_prob_NN-0_1.csv')
# all_prob_RandomForest = pd.read_csv('case_AREPS/ml_output/NA/NA_candidates_all_prob_RandomForest_2.csv')
# all_prob_svc_poly = pd.read_csv('case_AREPS/ml_output/NA/NA_candidates_all_prob_SVC-POLY-C_10.csv')
# all_prob_svc_rbf = pd.read_csv('case_AREPS/ml_output/NA/NA_candidates_all_prob_SVC-RBF-C_10.csv')

# files = [all_prob_adaboost, all_prob_GaussianProcess, all_prob_NN, all_prob_RandomForest, all_prob_svc_poly, all_prob_svc_rbf]

# for file in files:
#     print(file)
#     NA_candidates_all_ages_mean = file.groupby(['lon','lat'])['prob'].mean()
#     print(NA_candidates_all_ages_mean.shape)
#     NA_candidates_all_ages_mean.to_csv('case_AREPS/ml_output/NA/NA_candidates_all_mean_prob_'+name+'.csv', index=True)
#     print('Ok!')

# print('finish')


SA_all_positive_data = pd.read_csv('case_AREPS/ml_input/SA_positive_all_columns.csv')
SA_all_negative_data = pd.read_csv('case_AREPS/ml_input/SA_negative_all_columns.csv')
#NA_all_candidates_data = pd.read_csv('case_AREPS/coreg_output_clean/NA/deposit_candidates.csv')


feature_names = Utils.get_parameter('feature_names')

SA_positive_idx = SA_all_positive_data[feature_names].dropna().index
SA_negative_idx = SA_all_negative_data[feature_names].dropna().index
SA_candidates_idx = SA_all_candidates_data[feature_names].dropna().index


SA_positives_lat_lon = SA_all_positive_data.iloc[SA_positive_idx][['lat', 'lon', 'recon_lon', 'recon_lat', 'age']]
SA_negatives_lat_lon = SA_all_negative_data.iloc[SA_negative_idx][['lat', 'lon', 'recon_lon', 'recon_lat', 'age']]    

SA_positives_lat_lon['prob'] = 10
SA_negatives_lat_lon['prob'] = 20
SA_positives_lat_lon['method']='positives'
SA_negatives_lat_lon['method']='negatives'

df_2 = pd.DataFrame(columns=['lat', 'lon', 'recon_lon', 'recon_lat', 'age', 'prob', 'method']) 

df_2 = pd.concat([SA_positives_lat_lon, SA_negatives_lat_lon, df])

df_2.to_csv('case_AREPS/ml_output/SA/SA_candidates_pos_neg_all_prob_.csv', index=True)
